Don't query fields to compute display label/HFID
After the HFID and display label refactoring work, we started storing
the values for these properties as attributes in the database. We
therefore do not require any logic to try to compute them on the fly
and so we do not need to find out which fields to query for alongside
those that the user requested.

This change fixes cases where HFID and display label may not be
retrieved from the database when trying to access them for relationships
via GraphQL queries.