fix(backend): do not delete rels when previewing mutation object
PR #7508 introduced a regression where relationship constraints would
not rollback the transaction in case of failures due to the introduction
of a "preview object" executed outside of the transaction.
This would leave objects with nulled relationships where it would not
have been allowed.

Make sure those deletions are not processed when building this preview
object.
Also add more coverage to the unit test.

Fixes #7853

Signed-off-by: Fatih Acar <fatih@opsmill.com>