IFC-1955 - Update frontend to handle relationships in profiles (#7837)
* IFC-1952 Add generic/attribute relationships in profile schema

This change is relatively simple and just adds relationships of kind generic and kind attribute to profile schemas corresponding to node schemas having these relationships.

When we load a schema, we make sure to clear all the relationships, except the ones of kind PROFILE such as related_nodes, to create a brand new list of relationships which will account for deleted and new ones. This is very similar to what we do for templates already.

* IFC-1961 Add ability to clear out profile relationships on updates (#7680)

* IFC-1953 Apply relationships from profiles (#7660)

* start add profiles in relationships

* add plan

* changelog

* IFC-1952 Add generic/attribute relationships in profile schema

This change is relatively simple and just adds relationships of kind generic and kind attribute to profile schemas corresponding to node schemas having these relationships.

When we load a schema, we make sure to clear all the relationships, except the ones of kind PROFILE such as related_nodes, to create a brand new list of relationships which will account for deleted and new ones. This is very similar to what we do for templates already.

* IFC-1961 Add ability to clear out profile relationships on updates (#7680)

* IFC-1953 Apply relationships from profiles (#7660)

* Adapt code to new source metadata

* Sort import

* Test fix

* IFC-1954 Add events to apply profiles on attr/rel change (#7807)

* Prefetch relationships to avoid calling `fetch()`

* Fix `related_nodes` fetch in profile refresh task

* fix import

* remove logs

* IFC-1952 Add generic/attribute relationships in profile schema

This change is relatively simple and just adds relationships of kind generic and kind attribute to profile schemas corresponding to node schemas having these relationships.

When we load a schema, we make sure to clear all the relationships, except the ones of kind PROFILE such as related_nodes, to create a brand new list of relationships which will account for deleted and new ones. This is very similar to what we do for templates already.

* IFC-1961 Add ability to clear out profile relationships on updates (#7680)

* IFC-1953 Apply relationships from profiles (#7660)

* Adapt code to new source metadata

* Sort import

* Test fix

* IFC-1954 Add events to apply profiles on attr/rel change (#7807)

* Prefetch relationships to avoid calling `fetch()`

* Fix `related_nodes` fetch in profile refresh task

* extract metadata fragment

* ensure all values are from profile

* remove plan

* remove changelog

* remove logs

* remove old code

* Improve logic to apply relationships

* fix edge cases

* Applying profile should be triggered if priority changes

* Add docker integration tests

* Fix mypy errors

* Remove unused function

* Lint

---------

Co-authored-by: Guillaume Mazoyer <guillaume@opsmill.com>