fix(backend, tests): too many db connections in integration tests
Since the neo4j driver upgrade, we must ensure the neo4j driver is
properly closed and that a closed driver is not re-used.
We disabled singletons during DI to get that behavior.

A side effect is that during some integration tests that leverage local
workers, some tasks/flows depend on the injected database instance,
which would trigger new driver instances with new db connections.
Those instances would not get cleaned since it's not using a singleton
nor used within a context manager that would close the underlying db
driver.

Fix this by introducing a new class level fixture for the db instance
and use it within the injected dependency.

Signed-off-by: Fatih Acar <fatih@opsmill.com>